+++ src/main/java/com/metrostate/edu/booknow/models/Customer.java
@@ -1,28 +1,16 @@
 package com.metrostate.edu.booknow.models;
 
 public class Customer {
-    private int id;
     private String username;
     private String password;
-    private String phoneNumber;
 
     // Constructor
-    public Customer(int id, String username, String password, String phoneNumber) {
-        this.id = id;
+    public Customer(String username, String password) {
         this.username = username;
         this.password = password;
-        this.phoneNumber = phoneNumber;
     }
 
     // Getters and setters
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
     public String getUsername() {
         return username;
     }
@@ -37,13 +25,5 @@ public class Customer {
 
     public void setPassword(String password) {
         this.password = password;
-    }
-
-    public String getPhoneNumber() {
-        return phoneNumber;
-    }
-
-    public void setPhoneNumber(String phoneNumber) {
-        this.phoneNumber = phoneNumber;
     }
 }