+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # ChangeLog
 
+## [1.0.1] - -09/29/24 - Jimmie Xiong
+### Added
+- Implemented the `displayRestaurantDetails(Restaurant restaurant)` method in `BookNowController` for displaying detailed information about selected restaurants.
+- Added a new FXML file section for the restaurant details page.
+- Created the `RestaurantDetailsView` UI for displaying restaurant details, including menu, operating hours, and customer reviews.
+- Added methods to the `BookNowController` 
+
 ## [1.0.0] - 09/28/24 - Jimmie Xiong
 ### Added
 - Developed models: `Customer`, `Restaurant`, and `Reservation`.